@@ -1,52 +1,63 @@
 {
-    "name": "webpak-default-project",
-    "version": "1.0.0",
-    "description": "",
-    "main": "src/assets/js/index.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "build": "cross-env NODE_ENV=production node server/index.js",
-        "start": "nodemon --config nodemon.json",
-        "prod": "webpack --env.production=true --config config/webpack.config.js"
-    },
-    "author": "",
-    "license": "ISC",
-    "devDependencies": {
-        "ansi": "^0.3.1",
-        "autoprefixer": "^8.6.3",
-        "babel-core": "^6.26.3",
-        "babel-loader": "^7.1.4",
-        "babel-plugin-transform-custom-element-classes": "^0.1.0",
-        "babel-plugin-transform-runtime": "^6.23.0",
-        "babel-preset-env": "^1.7.0",
-        "chokidar": "^2.0.4",
-        "clean-webpack-plugin": "^0.1.19",
-        "compression-webpack-plugin": "^1.1.11",
-        "cross-env": "^5.2.0",
-        "css-loader": "^0.28.11",
-        "cssnano": "^4.0.0",
-        "eventsource-polyfill": "^0.9.6",
-        "express": "^4.16.3",
-        "express-static-gzip": "^0.3.2",
-        "file-loader": "^1.1.11",
-        "html-loader": "^0.5.5",
-        "html-webpack-plugin": "^3.2.0",
-        "image-webpack-loader": "^4.3.1",
-        "mini-css-extract-plugin": "^0.4.0",
-        "node-sass": "^4.9.0",
-        "nodemon": "^1.18.4",
-        "opn": "^5.3.0",
-        "optimize-css-assets-webpack-plugin": "^4.0.3",
-        "postcss-loader": "^2.1.5",
-        "sass-loader": "^7.0.3",
-        "style-loader": "^0.21.0",
-        "uglifyjs-webpack-plugin": "^1.2.7",
-        "webpack": "^4.12.0",
-        "webpack-cli": "^3.0.8",
-        "webpack-dev-middleware": "^3.1.3",
-        "webpack-hot-middleware": "^2.22.2"
-    },
-    "dependencies": {
-        "@webcomponents/custom-elements": "^1.1.2"
-    }
+  "name": "fdev-default",
+  "version": "0.0.1",
+  "description": "Default project for front-end development",
+  "main": "./server/main.ts",
+  "scripts": {
+    "start": "npm run build:live",
+    "build": "cross-env NODE_ENV=production ts-node ./server/main.ts",
+    "build:live": "cross-env NODE_ENV=development nodemon --config nodemon.json"
+  },
+  "keywords": [
+    "Front-end"
+  ],
+  "author": "StepanyanS",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/core": "^7.1.0",
+    "@babel/preset-env": "^7.1.0",
+    "@types/chokidar": "^1.7.5",
+    "@types/clean-webpack-plugin": "^0.1.2",
+    "@types/compression-webpack-plugin": "^0.4.2",
+    "@types/express": "^4.16.0",
+    "@types/html-webpack-plugin": "^3.2.0",
+    "@types/mini-css-extract-plugin": "^0.2.0",
+    "@types/node": "^10.10.1",
+    "@types/opn": "^5.1.0",
+    "@types/uglifyjs-webpack-plugin": "^1.1.0",
+    "@types/webpack": "^4.4.12",
+    "@types/webpack-dev-middleware": "^2.0.2",
+    "@types/webpack-dev-server": "^3.1.1",
+    "@types/webpack-hot-middleware": "^2.16.4",
+    "ansi": "^0.3.1",
+    "autoprefixer": "^9.1.5",
+    "babel-loader": "^8.0.2",
+    "chokidar": "^2.0.4",
+    "clean-webpack-plugin": "^0.1.19",
+    "compression-webpack-plugin": "^2.0.0",
+    "cross-env": "^5.2.0",
+    "css-loader": "^1.0.0",
+    "eventsource-polyfill": "^0.9.6",
+    "express": "^4.16.3",
+    "file-loader": "^2.0.0",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "image-webpack-loader": "^4.3.1",
+    "mini-css-extract-plugin": "^0.4.3",
+    "node-sass": "^4.9.3",
+    "nodemon": "^1.18.4",
+    "opn": "^5.4.0",
+    "postcss-loader": "^3.0.0",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.0",
+    "ts-node": "^7.0.1",
+    "typescript": "^3.0.3",
+    "uglifyjs-webpack-plugin": "^2.0.1",
+    "webpack": "^4.19.1",
+    "webpack-dev-middleware": "^3.3.0",
+    "webpack-hot-middleware": "^2.24.2"
+  },
+  "dependencies": {
+    "@webcomponents/custom-elements": "^1.2.0"
+  }
 }
